requirements.txt

pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart

# Vector Database and Embeddings
qdrant-client[fastembed]
langchain-huggingface
sentence-transformers
langchain_qdrant
unstructured[pdf]
unstructured_inference

# For GenAI
groq
langchain
langchain-text-splitters
langchain-community